Improve float serialization, add .NET 5 Half support.
Float serialization used the same serialization code as normal integers, which writes less bytes if the integer is of sufficiently small size. The problem however is that a float is basically *never* so small due to how the bits are laid out and this basically always resulted in a wasted extra byte vs just writing the 4/8 bytes of the float directly.

("basically never" is everything except 0/subnormals, which I'd say is not worth adding an extra byte to all other floats over).

The new code is also faster on Core because it writes the 4/8 bytes at once instead of 4/8 WriteByte/ReadByte() calls (see benchmarks in PR).

I also added support for the .NET 5 Half type.

Compiling for netcoreapp3.0 throws an "out of support" warning on the .NET 5 SDK so I changed the projects to target netcoreapp3.1 which is LTS instead.